refacor and add test to check failed test filter
