+++ crates/forge/Cargo.toml
@@ -62,6 +62,8 @@ tokio-util.workspace = true
 futures.workspace = true
 num-integer = "0.1.45"
 url.workspace = true
+mockall = "0.11.4"
+mockall_double = "0.3.0"
 
 [lib]
 name = "forge"
