+++ crates/forge/src/lib.rs
@@ -8,7 +8,7 @@ use camino::{Utf8Path, Utf8PathBuf};
 
 use futures::StreamExt;
 use running::{run_fuzz_test, run_test};
-use scarb_metadata::MetadataCommand;
+
 use tokio::sync::mpsc::{channel, Sender};
 
 use std::sync::Arc;
@@ -43,6 +43,7 @@ use test_collector::{FuzzerConfig, LinkedLibrary, RawForkConfig, RawForkParams, 
 
 pub mod pretty_printing;
 pub mod scarb;
+pub mod shared_cache;
 pub mod test_case_summary;
 pub mod test_filter;
 
@@ -53,7 +54,6 @@ mod test_crate_summary;
 mod test_execution_syscall_handler;
 
 pub const FUZZER_RUNS_DEFAULT: u32 = 256;
-pub const CACHE_DIR: &str = ".snfoundry_cache";
 
 static BUILTINS: Lazy<Vec<&str>> = Lazy::new(|| {
     vec![
@@ -149,14 +149,6 @@ pub enum CrateLocation {
     Lib,
     /// Crate in the `tests/` directory
     Tests,
-}
-
-pub fn cache_dir() -> Result<Utf8PathBuf> {
-    let scarb_metadata = MetadataCommand::new().inherit_stderr().exec()?;
-    let workspace_root = scarb_metadata.workspace.root.clone();
-    let cache_dir_path = workspace_root.join(CACHE_DIR);
-    std::fs::create_dir_all(&cache_dir_path)?;
-    Ok(cache_dir_path)
 }
 
 fn parse_fork_params(raw_fork_params: &RawForkParams) -> Result<ValidatedForkConfig> {
