+++ crates/forge/src/test_filter.rs
@@ -1,10 +1,10 @@
-use std::fs::File;
-use std::io::{BufRead, BufReader};
-
 use crate::collecting::{CompiledTestCrate, CompiledTestCrateRaw, ValidatedForkConfig};
-use crate::{cache_dir, TestCaseFilter};
-use anyhow::Result;
+use crate::TestCaseFilter;
+use mockall_double::double;
 use test_collector::TestCase;
+
+#[double]
+use crate::shared_cache::shared_cache;
 
 #[derive(Debug, PartialEq)]
 // Specifies what tests should be included
@@ -68,7 +68,6 @@ impl TestsFilter {
 
     pub(crate) fn filter_tests(&self, test_crate: CompiledTestCrateRaw) -> CompiledTestCrateRaw {
         let mut cases = test_crate.test_cases;
-
         cases = match &self.name_filter {
             NameFilter::All => cases,
             NameFilter::Match(filter) => cases
@@ -78,7 +77,7 @@ impl TestsFilter {
             NameFilter::ExactMatch(name) => {
                 cases.into_iter().filter(|tc| tc.name == *name).collect()
             }
-            NameFilter::RerunFailed => match Self::read_tests_failed_file() {
+            NameFilter::RerunFailed => match shared_cache::read_tests_failed_file() {
                 Ok(result) => cases
                     .into_iter()
                     .filter(|tc| result.iter().any(|name| name == &tc.name))
@@ -98,16 +97,6 @@ impl TestsFilter {
             ..test_crate
         }
     }
-
-    fn read_tests_failed_file() -> Result<Vec<String>> {
-        let tests_failed_path = cache_dir()?.join(".prev_tests_failed");
-        let file = File::open(tests_failed_path)?;
-        let buf = BufReader::new(file);
-        Ok(buf
-            .lines()
-            .map(|l| l.expect("Could not parse line"))
-            .collect())
-    }
 }
 
 impl TestCaseFilter for TestsFilter {
@@ -122,6 +111,7 @@ impl TestCaseFilter for TestsFilter {
 
 #[cfg(test)]
 mod tests {
+    use super::shared_cache;
     use crate::collecting::CompiledTestCrate;
     use crate::test_filter::TestsFilter;
     use crate::CrateLocation;
@@ -266,6 +256,26 @@ mod tests {
             TestsFilter::from_flags(Some("nonexistent".to_string()), false, false, false, false);
         let filtered = tests_filter.filter_tests(mocked_tests.clone());
         assert_eq!(filtered.test_cases, vec![]);
+
+        let tests_filter = TestsFilter::from_flags(Some(String::new()), false, false, false, true);
+
+        let ctx = shared_cache::read_tests_failed_file_context();
+        ctx.expect()
+            .returning(|| Ok(vec!["crate1::do_thing".to_string()]));
+
+        let filtered = tests_filter.filter_tests(mocked_tests.clone());
+
+        assert_eq!(
+            filtered.test_cases,
+            vec![TestCase {
+                name: "crate1::do_thing".to_string(),
+                available_gas: None,
+                ignored: false,
+                expected_result: ExpectedTestResult::Success,
+                fork_config: None,
+                fuzzer_config: None,
+            },]
+        );
 
         let tests_filter = TestsFilter::from_flags(Some(String::new()), false, false, false, false);
         let filtered = tests_filter.filter_tests(mocked_tests.clone());
